test(governance): give the Inventory suites one import door, and split the fixtures out
Two findings from the convention scan.

The mocks live in the shared harness, which diverges from the house rule that
vi.mock stays in each test file (modelProviderDrawerHarness.tsx states it). The
rule guards one hazard: an import above the harness that transitively reaches a
mocked module binds the real one, and the failure reads as a missing tRPC
provider. Keeping five copies of the mocks would restore the duplication the
split removed and push the largest suite back over the size limit, so the
harness now re-exports findNativeSelects and SAMPLE_CHOICE_KEY and no suite
imports anything but the harness and the test libraries. There is no line above
it for the hazard to sit on.

That left the harness at 315 non-blank lines, over the same 300 the split
existed to satisfy. CONNECTED_SOURCES moves to inventoryFixtures.ts — data
rather than setup, and imported by the harness, so it adds no ordering risk.
Harness 266, fixtures 59, largest suite 268.
