+++ platform/app/ee/governance/dashboard/pages/__tests__/inventoryUiRules.integration.test.tsx
@@ -15,11 +15,10 @@ import { screen, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { describe, expect, it } from "vitest";
 
-import { findNativeSelects } from "~/components/governance/filters";
-
 import {
   connectTools,
   emptyWithSamplesOff,
+  findNativeSelects,
   harness,
   openTab,
   renderScreen,
